+++ lib/spectre/helpers.rb
@@ -1,5 +1,3 @@
-require_relative '../spectre'
-
 require 'securerandom'
 require 'json'
 require 'date'